requirements.txt
flask>=2.0.1,<3.0.0
python-dotenv==1.0.0
sentence-transformers>=2.2.2
numpy>=1.21.0,<2.0.0
pyyaml>=6.0.1
python-Levenshtein>=0.21.1
torch>=2.0.0
transformers>=4.30.0
scikit-learn>=1.0.2
flask-session>=0.5.0 